requirements.txt
# Core
django==4.2.9
python-dotenv==1.0.1
str2bool==1.1

# UI 
django-admin-datta==1.0.10

# Tooling
django-dynamic-datatb==1.0.25
django-api-generator==1.0.17
django-debug-toolbar==4.2.0

# Deployment
whitenoise==6.6.0
gunicorn==21.2.0

# DB Layer
psycopg2-binary
mysqlclient

# Background Tasks
django-background-tasks-updated==1.2.8

# Summarization
beautifulsoup4
requests
tqdm
langchain
ollama
tiktoken
torch
fairscale
blobfile
